requirements.txt
numpy
pandas
matplotlib
scipy
jupyter
opencv-python
insightface
onnxruntime
pillow
psycopg2
python-dotenv